chore(vault): Moved and renamed SecretStore
